requirements.txt
enum-compat==0.0.2
python-registry==1.3.1
unicodecsv==0.14.1
